move Connection class
