+++ bundles/org.smarthomej.binding.amazonechocontrol/src/main/java/org/smarthomej/binding/amazonechocontrol/internal/handler/AccountHandler.java
@@ -55,11 +55,11 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.smarthomej.binding.amazonechocontrol.internal.AccountHandlerConfig;
 import org.smarthomej.binding.amazonechocontrol.internal.AccountServlet;
-import org.smarthomej.binding.amazonechocontrol.internal.Connection;
-import org.smarthomej.binding.amazonechocontrol.internal.ConnectionException;
 import org.smarthomej.binding.amazonechocontrol.internal.channelhandler.ChannelHandler;
 import org.smarthomej.binding.amazonechocontrol.internal.channelhandler.ChannelHandlerSendMessage;
 import org.smarthomej.binding.amazonechocontrol.internal.channelhandler.IAmazonThingHandler;
+import org.smarthomej.binding.amazonechocontrol.internal.connection.Connection;
+import org.smarthomej.binding.amazonechocontrol.internal.connection.ConnectionException;
 import org.smarthomej.binding.amazonechocontrol.internal.connection.LoginData;
 import org.smarthomej.binding.amazonechocontrol.internal.discovery.AmazonEchoDiscovery;
 import org.smarthomej.binding.amazonechocontrol.internal.discovery.SmartHomeDevicesDiscovery;
